First attempt with sliding tripod mount.
Created a tripod mount that slides onto the bottom of the housing. The lower piece has a screw hole for a typical tripod mount (1/4-20 UNC thread).

Also added small arms that are attached to the door and extend into the main body of the housing. Can use an M3 threaded heat set insert in these branches. Screws can then be pushed through holes in the side of the box. These enter the inserts and can be tightened to hold the door shut tightly.