Small modifications
Reoriented USB-C charging cap. Vertical gap in cap design of last model allowed water in. Rotated by 90 degrees, changed size so there is now a horizontal crevice when the two pieces are attached.

Added holes for threaded inserts in blocks on the door, also added holes in the side of the housing for screws to enter. These help align the screws and the inserts.